03-mcp/mcp-ai-dev-workflow/pets-api/pyproject.toml
[project]
name = "petstore-api"
version = "0.1.0"
description = "Swagger Petstore API implementation with Django REST Framework"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "Petstore Team", email = "apiteam@swagger.io"}
]

dependencies = [
    "django>=5.2,<6.0",
    "djangorestframework>=3.15.0",
    "djangorestframework-simplejwt>=5.4.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "psycopg2-binary>=2.9.10",
    "django-filter>=24.0",
    "drf-spectacular>=0.28.0",
    "pillow>=11.0.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.0",
    "gunicorn>=23.0.0",
    "whitenoise>=6.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.9.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.6.0",
    "factory-boy>=3.3.0",
    "faker>=33.0.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "django-stubs>=5.1.0",
    "djangorestframework-stubs>=3.15.0",
    "ipython>=8.30.0",
    "ipdb>=0.13.13",
    "django-extensions>=3.2.0",
    "werkzeug>=3.0.0",
    "pygments>=2.18.0",
    "flake8>=7.0.0",
    "isort>=5.13.0",
    "bandit>=1.7.0",
    "pylint>=3.0.0",
    "pylint-django>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "migrations",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["config", "apps", "common"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py", "tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--no-cov-on-fail",
]
testpaths = ["src"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__init__.py",
    "*/config/wsgi.py",
    "*/config/asgi.py",
    "*/manage.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
known_django = ["django"]
known_first_party = ["config", "pets", "store", "users"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = ["migrations", ".venv", "venv", "build", "dist"]

[tool.pylint.main]
load-plugins = ["pylint_django"]
django-settings-module = "config.settings.development"
ignore = ["migrations", ".venv", "venv"]
ignore-patterns = ["test_.*?.py"]

[tool.pylint.format]
max-line-length = 100

[tool.pylint.messages_control]
disable = [
    "C0103",  # Invalid name (Django uses lowercase for module names)
    "C0114",  # Missing module docstring
    "C0115",  # Missing class docstring
    "C0116",  # Missing function docstring
    "R0903",  # Too few public methods (Django models often have few methods)
    "W0212",  # Protected member access (Django uses _meta)
    "E1101",  # Instance has no member (Django models dynamic attributes)
]

[tool.pylint.design]
max-attributes = 10
max-args = 8

[tool.bandit]
exclude_dirs = ["/tests", "/migrations", "/.venv"]
skips = ["B101", "B601"]
